refactor: 상풀 상세 페이지 레이아웃 하단 공통 영역 분리 #31
